requirements.txt
pip==19.0.3
pyzmq
